requirements.txt
python-dotenv
pycord
gspread